Implement conversion statistics
Added counters for notebooks, pages, attachments and images.  NOTE will add a further counter when commence adding features for pre and post processing of notes to count the number of processing edits made.

Added timer for time taken for processing.

Changed getters to properties in NSXFile and added properties for new counters.