Implement handling of charts
Chart data is retrieved from note station html that does not parse through pandoc.  Chart is generated using matplotlib and an image is inserted into the note.  Below the image is a link to a csv file of the chart data, and below that is a table of the chart data.