Implement meta data header in pre-processing using class NSMetaDataGenerator
In pre-processing meta data is added in <meta> blocks to the note html before conversion or export.  A text version is generated alongside and is available for use in post processing where YAML/plain text header has to be entered manually.

Added -s option to pandoc conversion to have yaml header auto generated during conversion id using pandoc markdown export format (not yet an option but will be added but the -s is already there now)

Fixed - obsidian quick setting was not using obsidian export format.. it uses gfm in the background so I had not changed it from the default which is gfm.  Noticed when adding -raw_html option to obsidian export format